requirements.txt
azure-cognitiveservices-speech
websocket-client
requests
pilk
pydub